Moved CPU and GPU identification classes,
from the shared library to the platform library.
