+++ libraries/gpu-gl-common/CMakeLists.txt
@@ -1,6 +1,6 @@
 set(TARGET_NAME gpu-gl-common)
 setup_hifi_library(Concurrent)
-link_hifi_libraries(shared gl gpu shaders)
+link_hifi_libraries(shared platform gl gpu shaders)
 GroupSources("src")
 
 set(OpenGL_GL_PREFERENCE "GLVND")
