+++ libraries/platform/src/GPUIdent.cpp
