+++ libraries/platform/src/platform/backend/AndroidPlatform.cpp
@@ -8,7 +8,7 @@
 
 #include "AndroidPlatform.h"
 #include "../PlatformKeys.h"
-#include <GPUIdent.h>
+#include "../../GPUIdent.h"
 #include <QSysInfo>
 #include <QtCore/QtGlobal>
