+++ libraries/platform/src/platform/backend/LinuxPlatform.cpp
@@ -11,11 +11,11 @@
 
 #include <thread>
 #include <string>
-#include <CPUIdent.h>
+#include "../../CPUIdent.h"
 
 #include <QtCore/QtGlobal>
 
-#include <GPUIdent.h>
+#include "../../GPUIdent.h"
 #include <QSysInfo>
 
 using namespace platform;
@@ -50,7 +50,7 @@ void LinuxInstance::enumerateMemory() {
 }
 
 void LinuxInstance::enumerateComputer(){
-    
+
     _computer[keys::computer::OS] = keys::computer::OS_LINUX;
     _computer[keys::computer::vendor] = "";
     _computer[keys::computer::model] = "";
